requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==24.8.1
babel==2.17.0
billiard==4.2.1
bleach==6.2.0
celery==5.5.1
certifi==2025.6.15
cffi==1.17.1
channels==4.2.0
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.2
daphne==4.1.2
diff-match-patch==20241021
dj-database-url==3.0.0
Django==5.1.7
django-chunked-upload==2.0.0
django-extensions==4.1
django-filter==25.1
django-hordak==2.0.0
django-import-export==4.3.8
django-js-asset==3.1.2
django-money==3.5.4
django-mptt==0.17.0
django-vite==3.1.0
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
drf-chunked-upload==0.6.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
kombu==5.5.2
Markdown==3.8
prompt_toolkit==3.0.50
py-moneyed==3.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.19.1
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
requests==2.32.4
service-identity==24.2.0
setuptools==78.0.2
six==1.17.0
sqlparse==0.5.3
tablib==3.8.0
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
zope.interface==7.2
